+++ inst/app/init.R
@@ -8,23 +8,28 @@
 # options(shiny.error = recover)
 # options(warn = 2)
 
-if (getOption("radiant.shinyFiles", FALSE)) {
+if (isTRUE(getOption("radiant.shinyFiles", FALSE))) {
   if (isTRUE(Sys.getenv("RSTUDIO") == "") && isTRUE(Sys.getenv("SHINY_PORT") != "")) {
     ## Users not on Rstudio will only get access to pre-specified volumes
     sf_volumes <- getOption("radiant.sf_volumes", "")
   } else {
-    sf_volumes <- getOption("radiant.launch_dir")
-    if (getOption("radiant.project_dir", "") != "") {
-      sf_volumes <- unique(getOption("radiant.project_dir"), sf_volumes)
+    if (getOption("radiant.project_dir", "") == "") {
+      sf_volumes <- getOption("radiant.launch_dir") %>% {set_names(., basename(.))}
+    } else {
+      sf_volumes <- getOption("radiant.project_dir") %>% {set_names(., basename(.))}
     }
     home <- radiant.data::find_home()
     if (home != sf_volumes) {
-      sf_volumes <- c(sf_volumes, home) %>% set_names(c(basename(sf_volumes), "Home"))
+      sf_volumes <- c(sf_volumes, home) %>% set_names(c(names(sf_volumes), "Home"))
     } else {
       sf_volumes <- c(Home = home)
     }
     if (sum(nzchar(getOption("radiant.sf_volumes", ""))) > 0) {
       sf_volumes <- getOption("radiant.sf_volumes") %>% {c(sf_volumes, .[!. %in% sf_volumes])}
+    }
+    missing_names <- is.na(names(sf_volumes))
+    if (sum(missing_names) > 0) {
+      sf_volumes[missing_names] <- basename(sf_volumes[missing_names])
     }
   }
 }